Added Custom Styles and Tooltip
