+++ community-chatbot/components/agent/AppTour/styles.ts
@@ -1,0 +1,14 @@
+export const getJoyrideStyles = () => ({
+    options: {
+        overlayColor: 'rgba(0, 0, 0, 0.6)',
+        primaryColor: '#2563eb',
+        zIndex: 10000,
+    },
+    spotlight: {
+        borderRadius: '1rem',
+    },
+    tooltip: {
+        backgroundColor: 'transparent',
+        padding: 0,
+    },
+});
